+++ lib/src/e2ee.worker/e2ee.keyhandler.dart
@@ -1,11 +1,14 @@
 import 'dart:async';
-import 'dart:html';
 import 'dart:js_util' as jsutil;
 import 'dart:typed_data';
+
+import 'package:web/web.dart' as web;
 
 import 'crypto.dart' as crypto;
 import 'e2ee.logger.dart';
 import 'e2ee.utils.dart';
+
+const KEYRING_SIZE = 16;
 
 class KeyOptions {
   KeyOptions({
@@ -14,12 +17,16 @@ class KeyOptions {
     required this.ratchetWindowSize,
     this.uncryptedMagicBytes,
     this.failureTolerance = -1,
+    this.keyRingSze = KEYRING_SIZE,
+    this.discardFrameWhenCryptorNotReady = false,
   });
   bool sharedKey;
   Uint8List ratchetSalt;
   int ratchetWindowSize = 0;
   int failureTolerance;
   Uint8List? uncryptedMagicBytes;
+  int keyRingSze;
+  bool discardFrameWhenCryptorNotReady;
 
   @override
   String toString() {
@@ -29,7 +36,7 @@ class KeyOptions {
 
 class KeyProvider {
   KeyProvider(this.worker, this.id, this.keyProviderOptions);
-  final DedicatedWorkerGlobalScope worker;
+  final web.DedicatedWorkerGlobalScope worker;
   final String id;
   final KeyOptions keyProviderOptions;
   var participantKeys = <String, ParticipantKeyHandler>{};
@@ -76,12 +83,10 @@ class KeyProvider {
   }
 }
 
-const KEYRING_SIZE = 16;
-
 class KeySet {
   KeySet(this.material, this.encryptionKey);
-  CryptoKey material;
-  CryptoKey encryptionKey;
+  web.CryptoKey material;
+  web.CryptoKey encryptionKey;
 }
 
 class ParticipantKeyHandler {
@@ -89,10 +94,15 @@ class ParticipantKeyHandler {
     required this.worker,
     required this.keyOptions,
     required this.participantIdentity,
-  });
+  }) {
+    if (keyOptions.keyRingSze <= 0 || keyOptions.keyRingSze > 255) {
+      throw Exception('Invalid key ring size');
+    }
+    cryptoKeyRing = List.filled(keyOptions.keyRingSze, null);
+  }
   int currentKeyIndex = 0;
 
-  List<KeySet?> cryptoKeyRing = List.filled(KEYRING_SIZE, null);
+  late List<KeySet?> cryptoKeyRing;
 
   bool _hasValidKey = false;
 
@@ -100,7 +110,7 @@ class ParticipantKeyHandler {
 
   final KeyOptions keyOptions;
 
-  final DedicatedWorkerGlobalScope worker;
+  final web.DedicatedWorkerGlobalScope worker;
 
   final String participantIdentity;
 
@@ -157,8 +167,8 @@ class ParticipantKeyHandler {
     return newKey;
   }
 
-  Future<CryptoKey> ratchetMaterial(
-      CryptoKey currentMaterial, ByteBuffer newKeyBuffer) async {
+  Future<web.CryptoKey> ratchetMaterial(
+      web.CryptoKey currentMaterial, ByteBuffer newKeyBuffer) async {
     var newMaterial = await jsutil.promiseToFuture(crypto.importKey(
       'raw',
       newKeyBuffer,
@@ -194,14 +204,14 @@ class ParticipantKeyHandler {
 
   /// Derives a set of keys from the master key.
   /// See https://tools.ietf.org/html/draft-omara-sframe-00#section-4.3.1
-  Future<KeySet> deriveKeys(CryptoKey material, Uint8List salt) async {
+  Future<KeySet> deriveKeys(web.CryptoKey material, Uint8List salt) async {
     var algorithmOptions =
         getAlgoOptions((material.algorithm as crypto.Algorithm).name, salt);
 
     // https://developer.mozilla.org/en-US/docs/Web/API/SubtleCrypto/deriveKey#HKDF
     // https://developer.mozilla.org/en-US/docs/Web/API/HkdfParams
     var encryptionKey =
-        await jsutil.promiseToFuture<CryptoKey>(crypto.deriveKey(
+        await jsutil.promiseToFuture<web.CryptoKey>(crypto.deriveKey(
       jsutil.jsify(algorithmOptions),
       material,
       jsutil.jsify({'name': 'AES-GCM', 'length': 128}),
@@ -215,7 +225,7 @@ class ParticipantKeyHandler {
   /// Ratchets a key. See
   /// https://tools.ietf.org/html/draft-omara-sframe-00#section-4.3.5.1
 
-  Future<Uint8List> ratchet(CryptoKey material, Uint8List salt) async {
+  Future<Uint8List> ratchet(web.CryptoKey material, Uint8List salt) async {
     var algorithmOptions = getAlgoOptions('PBKDF2', salt);
 
     // https://developer.mozilla.org/en-US/docs/Web/API/SubtleCrypto/deriveBits