+++ lib/src/media_recorder_impl.dart
@@ -1,12 +1,13 @@
 import 'dart:async';
-import 'dart:html' as html;
 import 'dart:js' as js;
+import 'dart:js_interop';
+import 'package:web/web.dart' as web;
 import 'package:webrtc_interface/webrtc_interface.dart';
 
 import 'media_stream_impl.dart';
 
 class MediaRecorderWeb extends MediaRecorder {
-  late html.MediaRecorder _recorder;
+  late web.MediaRecorder _recorder;
   late Completer<String> _completer;
 
   @override
@@ -28,30 +29,38 @@ class MediaRecorderWeb extends MediaRecorder {
     int timeSlice = 1000,
   }) {
     var _native = stream as MediaStreamWeb;
-    _recorder = html.MediaRecorder(_native.jsStream, {'mimeType': mimeType});
+    _recorder = web.MediaRecorder(
+        _native.jsStream, web.MediaRecorderOptions(mimeType: mimeType));
     if (onDataChunk == null) {
-      var _chunks = <html.Blob>[];
+      var _chunks = <web.Blob>[];
       _completer = Completer<String>();
-      _recorder.addEventListener('dataavailable', (html.Event event) {
-        final html.Blob blob = js.JsObject.fromBrowserObject(event)['data'];
-        if (blob.size > 0) {
-          _chunks.add(blob);
-        }
-        if (_recorder.state == 'inactive') {
-          final blob = html.Blob(_chunks, mimeType);
-          _completer.complete(html.Url.createObjectUrlFromBlob(blob));
-        }
-      });
-      _recorder.onError.listen((error) {
-        _completer.completeError(error);
-      });
+      _recorder.addEventListener(
+          'dataavailable',
+          (web.Event event) {
+            final web.Blob blob = js.JsObject.fromBrowserObject(event)['data'];
+            if (blob.size > 0) {
+              _chunks.add(blob);
+            }
+            if (_recorder.state == 'inactive') {
+              final blob =
+                  web.Blob(_chunks.toJS, web.BlobPropertyBag(type: mimeType));
+              _completer.complete(web.URL.createObjectURL(blob));
+            }
+          }.toJS);
+      _recorder.addEventListener(
+          'error',
+          (JSAny error) {
+            _completer.completeError(error);
+          }.toJS);
     } else {
-      _recorder.addEventListener('dataavailable', (html.Event event) {
-        onDataChunk(
-          js.JsObject.fromBrowserObject(event)['data'],
-          _recorder.state == 'inactive',
-        );
-      });
+      _recorder.addEventListener(
+          'dataavailable',
+          (web.Event event) {
+            onDataChunk(
+              js.JsObject.fromBrowserObject(event)['data'],
+              _recorder.state == 'inactive',
+            );
+          }.toJS);
     }
     _recorder.start(timeSlice);
   }